+++ test/concurrent_limiter_test.exs
@@ -9,12 +9,31 @@ defmodule ConcurrentLimiterTest do
   test "limited to one" do
     name = "l1"
     ConcurrentLimiter.new(name, 1, 0, max_retries: 0)
-    endless = fn() -> :timer.sleep(10000) end
-    spawn(fn() -> ConcurrentLimiter.limit(name, endless) end)
+    endless = fn -> :timer.sleep(10_000) end
+    spawn(fn -> ConcurrentLimiter.limit(name, endless) end)
     :timer.sleep(5)
     {:error, :overload} = ConcurrentLimiter.limit(name, endless)
     {:error, :overload} = ConcurrentLimiter.limit(name, endless)
     {:error, :overload} = ConcurrentLimiter.limit(name, endless)
+  end
+
+  test "decrements correctly when current pid exits" do
+    name = "l1crash"
+    ConcurrentLimiter.new(name, 1, 0, max_retries: 0)
+    endless = fn -> :timer.sleep(100) end
+
+    pid =
+      spawn(fn ->
+        ConcurrentLimiter.limit(name, endless)
+      end)
+
+    # let some time for spawn to execute
+    :timer.sleep(5)
+    {:error, :overload} = ConcurrentLimiter.limit(name, endless)
+    Process.exit(pid, :kill)
+    # let some time for exit to execute
+    :timer.sleep(5)
+    :ok = ConcurrentLimiter.limit(name, fn -> :ok end)
   end
 
   test "limiter is atomic" do