+++ test/samples/limiter.exs
@@ -1,24 +1,13 @@
 infinite = 1_000_000_000_000_000_000_000_000_000_000_000_000_000_000_000
 ConcurrentLimiter.new(:bench, infinite, 0)
-ConcurrentLimiter.new(:bench_s, infinite, 0, ets: ConcurrentLimiterTest)
-
-concurrent = [{:read_concurrency, true}, {:write_concurrency, true}]
-
-ConcurrentLimiter.new(:bench_rw, infinite, 0)
-ConcurrentLimiter.new(:bench_s_rw, infinite, 0, ets: ConcurrentLimiterTest, ets_opts: concurrent)
+ConcurrentLimiter.new(:bench_no_sentinel, infinite, 0, sentinel: false)
 
 single = %{
-  "ConcurrentLimiter.limit/2" => fn ->
+  "ConcurrentLimiter.limit/2 (with sentinels)" => fn ->
     ConcurrentLimiter.limit(:bench, fn -> :ok end)
   end,
-  "ConcurrentLimiter.limit/2 with concurrency" => fn ->
-    ConcurrentLimiter.limit(:bench_rw, fn -> :ok end)
-  end,
-  "ConcurrentLimiter:limit/2 with shared ets" => fn ->
-    ConcurrentLimiter.limit(:bench_s, fn -> :ok end)
-  end,
-  "ConcurrentLimiter:limit/2 with shared ets and concurrency" => fn ->
-    ConcurrentLimiter.limit(:bench_s_rw, fn -> :ok end)
+  "ConcurrentLimiter.limit/2 (without sentinels)" => fn ->
+    ConcurrentLimiter.limit(:bench_no_sentinel, fn -> :ok end)
   end
 }